fix compile warning
# message auto-generated for no-merge-commit merge:
merge main into master

fix compile warning

Created-by: msprofDeveloper

Match-id-db079c66c941243cc4a69b9ac4e705723b20448c

+++ collector/dvvp/message/prof_json_config.cpp
@@ -32,7 +32,7 @@ const std::vector<std::string> aclCfgName = {"dvpp_freq", "host_sys_usage_freq",
 
 inline bool isValueInArray(const std::vector<std::string>& arr, std::string value)
 {
-    for (int i = 0; i < arr.size(); i++) {
+    for (size_t i = 0; i < arr.size(); i++) {
         if (arr[i] == value) {
             return true;
         }
